+++ tau/field.py
@@ -1,4 +1,5 @@
 import math
+from .app import fields, objects
 from .vector import Vector
 
 
@@ -18,7 +19,12 @@ class Field(object):
         self.object = object
         self._x = object.x
         self._y = object.y
-        self._app = None
+
+        fields.append(self)
+
+    def __del__(self):
+        objects.remove(self)
+        del self
 
     def radius(self, other):
         """Calculate the radius of the field.
@@ -33,14 +39,10 @@ class Field(object):
         """
         return math.sqrt((other.x - self._x) ** 2 + (other.y - self._y) ** 2)
 
-    def update(self, dt):
+    async def update(self):
         """Called every time a second passes in the simulation.
 
         This should be overridden by child classes to apply forces to objects.
-
-        Args:
-            dt: The floored (rounded down) amount of whole seconds that has
-                elasped between frame refreshes. This value is generally 0.
         """
         pass
 
@@ -64,12 +66,11 @@ class GravitationalField(Field):
         super().__init__(object)
         self._mass = object.mass
 
-    def update(self, dt):
-        for object in self._app.objects:
+    async def update(self, dt):
+        for object in objects:
             if object != self.object:
                 gf = (self.GRAVITATIONAL_CONSTANT * self._mass
                       * object.mass) / (self.radius(object) ** 2)
-                print(self.angle_between(object))
                 force = Vector(gf * math.sin(self.angle_between(object)),
                                gf * math.cos(self.angle_between(object)))
                 object.apply_force(force)
